requirements.txt
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dell-recovery==0.0.0
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
feedparser==5.2.1
filelock==3.3.2
future==0.18.2
guacamole==0.9.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
Jinja2==2.10.1
jsonschema==3.2.0
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
padme==1.1.1
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
pipenv==2021.11.9
platformdirs==2.4.0
progressbar==2.5
protobuf==3.6.1
psycopg2==2.9.1
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.2
system-service==0.3
systemd-python==234
texttable==1.6.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-devhelp==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-zotero==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
whitenoise==5.3.0
xdiagnose==3.8.9
xkit==0.0.0
XlsxWriter==1.1.2
zipp==1.0.0
